consumer/x/interoracle/keeper/grpc_query_oracle_info.go
package keeper

import (
	"context"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/maingockien01/consumer/x/interoracle/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) OracleInfo(c context.Context, req *types.QueryGetOracleInfoRequest) (*types.QueryGetOracleInfoResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(c)

	val, found := k.GetOracleInfo(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetOracleInfoResponse{OracleInfo: val}, nil
}
